+++ files/gitlab-cookbooks/gitlab/attributes/default.rb
@@ -293,6 +293,7 @@ default['gitlab']['gitlab-rails']['db_fdw'] = nil
 
 default['gitlab']['gitlab-rails']['redis_host'] = "127.0.0.1"
 default['gitlab']['gitlab-rails']['redis_port'] = nil
+default['gitlab']['gitlab-rails']['redis_ssl'] = false
 default['gitlab']['gitlab-rails']['redis_password'] = nil
 default['gitlab']['gitlab-rails']['redis_socket'] = "/var/opt/gitlab/redis/redis.socket"
 default['gitlab']['gitlab-rails']['redis_sentinels'] = []