+++ files/gitlab-cookbooks/gitlab/libraries/redis_helper.rb
@@ -5,18 +5,15 @@ class RedisHelper
     @node = node
   end
 
-  def redis_url(workhorse = false)
+  def redis_url
     gitlab_rails = @node['gitlab']['gitlab-rails']
 
     if gitlab_rails['redis_socket']
       uri = URI('unix:/')
       uri.path = gitlab_rails['redis_socket']
-    elsif workhorse
-      uri = URI('tcp:/')
-      uri.host = gitlab_rails['redis_host']
-      uri.port = gitlab_rails['redis_port']
     else
-      uri = URI::Redis.parse('redis:/')
+      scheme = gitlab_rails['redis_ssl'] ? 'rediss:/' : 'redis:/'
+      uri = URI(scheme)
       uri.host = gitlab_rails['redis_host']
       uri.port = gitlab_rails['redis_port']
       uri.password = gitlab_rails['redis_password']